--- src/main/java/com/example/rabbitmq/RabbitConfig.java
@@ -1,21 +1,0 @@
-package com.example.rabbitmq;
-
-import org.springframework.amqp.core.Queue;
-import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class RabbitConfig {
-    public static final String QUEUE = "email";
-
-    @Bean
-    public Queue queue(){
-        return new Queue(QUEUE, true);
-    }
-
-    @Bean
-    public Jackson2JsonMessageConverter messageConverter(){
-        return new Jackson2JsonMessageConverter();
-    }
-}
