+++ src/main/java/com/example/rabbitmq/rabbitmail/MailProducer.java
@@ -1,5 +1,7 @@
-package com.example.rabbitmq;
+package com.example.rabbitmq.rabbitmail;
 
+import com.example.rabbitmq.config.RabbitConfig;
+import com.example.rabbitmq.dto.RequestDTO;
 import lombok.AllArgsConstructor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
